feat: update embassy to latest (cf6a6cd)
Migrate all boards and shared crates from embassy rev 1d2f0ad/0ac791e
to cf6a6cd (latest HEAD). Key breaking changes addressed:

- I2c/Spi/Uart/ExtiInput constructors: reordered args, added DMA and
  EXTI interrupt bindings via _irq parameter
- bind_interrupts! macros: added explicit DMA channel and EXTI entries
  on every board
- Generic type updates: I2c<Async> -> I2c<'d, Async, mode::Master>,
  Spi<Async> -> Spi<'d, Async, mode::Master>,
  ExtiInput<'d> -> ExtiInput<'d, Async>
- embedded-io-async bumped 0.6 -> 0.7
- dual-bank feature added for STM32G4 chips
- NoopRawMutex replaced with CriticalSectionRawMutex where Sync required
- ADC API changes in trafag-pressure (AdcConfig, DMA irq bindings)
- PwmPin::new_ch* -> PwmPin::new, max_duty returns u32
- embassy-executor: arch-cortex-m -> platform-cortex-m, spawn API update
- Vendored postcard-rpc updated for embassy-usb 0.6, embassy-sync 0.8,
  embassy-executor 0.10, embedded-io-async 0.7

+++ fw-power-board/app/src/buzzer.rs
@@ -74,8 +74,8 @@ pub enum Step {
 /// *Global* loudness is still handled by `set_volume`.
 pub struct Buzzer<Tim: GeneralInstance4Channel> {
     pwm: SimplePwm<'static, Tim>,
-    max_duty: u16,
-    duty: u16,
+    max_duty: u32,
+    duty: u32,
     /// 0.0–100.0 %
     volume: f32,
 }
@@ -95,7 +95,7 @@ impl<Tim: GeneralInstance4Channel> Buzzer<Tim> {
 
     pub fn set_volume(&mut self, vol: f32) {
         self.volume = vol.min(100.0);
-        let duty = (self.max_duty as f32 * (self.volume / 100.0)).round() as u16;
+        let duty = (self.max_duty as f32 * (self.volume / 100.0)).round() as u32;
         self.duty = duty;
     }
     pub fn set_tone(&mut self, f: Hertz) {
